Map_Merging/Product/Multi_Map_Stitcher/requirements.txt
opencv-contrib-python==4.5.3.56
imutils==0.5.4
matplotlib==3.1.2
loguru==0.5.3
Pillow==8.2.0
tensorboard==2.6.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.6.0
tensorflow-estimator==2.6.0
numpy==1.19.5
keras==2.6.0
Keras-Preprocessing==1.1.2
caer==1.9.8
